Feature/place order (#22)
* feat(order-service): refactor order placement and entity handling

- Refactored `OrderService.placeOrder(OrderApiDto)` to:
  * Properly initialize `Order` entity and set `customer` & `orderDate`
  * Null-check `items` and handle empty lists
  * Correctly map `OrderItemApiDto` -> `OrderItem` using `OrderMapper`
  * Maintain bidirectional relationship with `order.addItem(item)` to ensure `order_id` is set
  * Decrement book stock and save updated `Book` entities

- Updated `OrderMapper`:
  * Added enum mapping between `OrderApiDto.StatusEnum` and `Order.OrderStatus`
  * Ensured `LocalDateTime` <-> `OffsetDateTime` conversion helpers
  * Avoided using builder for `Order` to prevent unknown property issues (`items` and `status`)

- Adjusted DTOs and OpenAPI contract:
  * Accept `customer.id` as Integer, convert to entity reference
  * Support multiple items with quantities per order

- Modified entity handling for `Order`, `OrderItem`, and `Book`:
  * Ensured `order_item` properly references parent `Order` (`order_id`) to fix SQL constraint issues
  * Corrected stock decrement logic for `Book` and total price calculation in `Order`
  * Enabled API responses to include proper `items` and `status` values

- Fixes:
  * `NullPointerException` on `orderApiDto.getItems()`
  * SQL constraint violations (`NULL not allowed for column ORDER_ID`)
  * Empty `items` and `status` fields in API responses

* ref: get all ids back to long instead of integer

* test: adapt tests to the new placeOrder

---------

+++ src/main/java/com/book/store/Entity/Book.java
@@ -12,12 +12,10 @@ import lombok.NoArgsConstructor;
 public class Book {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String author;
     private String title;
     private String description;
     private int quantity;
     private float price;
-
-    
 }
